// power-the-polls-form/src/components/address-input/SmartyStreets.d.ts
export declare type StateCode = ( 'AK' | 'AL' | 'AR' | 'AZ' | 'CA' | 'CO' | 'CT' | 'DC' | 'DE' | 'FL' | 'GA' | 'GU' | 'HI' | 'IA' | 'ID' | 'IL' | 'IN' | 'KS' | 'KY' | 'LA' | 'MA' | 'MD' | 'ME' | 'MI' | 'MN' | 'MO' | 'MS' | 'MT' | 'NC' | 'ND' | 'NE' | 'NH' | 'NJ' | 'NM' | 'NV' | 'NY' | 'OH' | 'OK' | 'OR' | 'PA' | 'PR' | 'RI' | 'SC' | 'SD' | 'TN' | 'TX' | 'UT' | 'VA' | 'VT' | 'WA' | 'WI' | 'WV' | 'WY' );
export interface GeoCoordinates {
   latitude?: number;
   longitude?: number;
}
export declare type GeoPrecision = "Unknown" | "None" | "State" | "SolutionArea" | "City" | "Zip5" | "Zip6" | "Zip7" | "Zip8" | "Zip9" | "Structure";
export interface GeoCoordinatesWithPrecision extends GeoCoordinates {
   precision?: GeoPrecision;
}
export declare namespace USStreetAddress {
   type AddressMatchType = "strict" | "range" | "invalid";
   interface QueryParamsItem {
      input_id?: string;
      street?: string;
      street2?: string;
      secondary?: string;
      city?: string;
      state?: string;
      zipcode?: string;
      lastline?: string;
      addressee?: string;
      urbanization?: string;
      candidates?: number;
      match?: AddressMatchType;
   }
   type QueryParams = QueryParamsItem | QueryParamsItem[];
   interface Components {
      urbanization?: string;
      primary_number?: string;
      street_name?: string;
      street_predirection?: string;
      street_postdirection?: string;
      street_suffix?: string;
      secondary_number?: string;
      secondary_designator?: string;
      extra_secondary_number?: string;
      extra_secondary_designator?: string;
      pmb_designator?: string;
      pmb_number?: string;
      city_name?: string;
      default_city_name?: string;
      state_abbreviation?: StateCode;
      zipcode?: string;
      plus4_code?: string;
      delivery_point?: string;
      delivery_point_check_digit?: string;
   }
   type RecordType = "F" | "G" | "H" | "P" | "R" | "S";
   type ZipType = "Unique" | "Military" | "POBox" | "Standard";
   type ResidentialDeliveryIndicator = "Residential" | "Commercial";
   type eLOTSoreOrder = "A" | "D";
   type CoordinatesPrecision = GeoPrecision;
   interface Metadata {
      record_type?: RecordType;
      zip_type?: ZipType;
      county_fips?: string;
      county_name?: string;
      carrier_route?: string;
      congressional_district?: string;
      building_default_indicator?: string;
      rdi?: ResidentialDeliveryIndicator;
      elot_sequence?: string;
      elot_sort?: eLOTSoreOrder;
      latitude?: number;
      longitude?: number;
      precision?: CoordinatesPrecision;
      time_zone?: string;
      utc_offset?: number;
      dst?: boolean;
   }
   type YesOrNo = "Y" | "N";
   type DVPMatchCode = "Y" | "N" | "S" | "D";
   type LACSLinkCode = "A" | "00" | "09" | "14" | "92";
   type LACSLinkIndicator = "Y" | "S" | "N" | "F";
   interface Analysis {
      dpv_match_code?: DVPMatchCode;
      dpv_footnotes?: string;
      dpv_cmra?: YesOrNo;
      dpv_vacant?: YesOrNo;
      active?: YesOrNo;
      ews_match?: boolean;
      footnotes?: string;
      lacslink_code?: LACSLinkCode;
      lacslink_indicator?: LACSLinkIndicator;
      suitelink_match?: boolean;
   }
   interface QueryResultItem {
      input_id?: string;
      input_index?: number;
      candidate_index?: number;
      addressee?: string;
      delivery_line_1?: string;
      delivery_line_2?: string;
      last_line?: string;
      delivery_point_barcode?: string;
      components?: Components;
      metadata?: Metadata;
      analysis?: Analysis;
   }
   type QueryResult = QueryResultItem[];
}
export declare namespace USAutocomplete {
   interface QueryParams {
      prefix: string;
      suggestions?: number;
      city_filter?: string;
      state_filter?: string;
      prefer?: string;
      prefer_ratio?: number;
      geolocate?: boolean;
      geolocate_precision?: string;
   }
   interface Suggestion {
      readonly text?: string;
      readonly street_line?: string;
      readonly city?: string;
      readonly state?: string;
      readonly zipcode?: string;
      readonly secondary?: string;
      readonly entries: number;
   }
   interface QueryResult {
      readonly suggestions: Suggestion[];
   }
}